[hm] Servo Then Fan
FanByServo mode works in a clunky fashion. PID algorithm has no idea of
the secondary effect of the fan.
In this mode Fan and servo operate in a serial fashion. First servo
(0-90), then they overlap ( servo 91-100, fan 0-10), then fan continues
on to 190.

Manual Mode can also use the 0-190 when the ServoThenFan mode is
selected.

Also did some changes to setOutputflags and storeOutputFlags so only one
mode of operation can be in effect.